Editor: Support Home and End keys. Handle "super" keys.
Home key will go to beginning of line, or if already there, then to the
first non-space character on the line.

End key will go to end of line (one past the last character), or if alrady there, to the last character on the line.

Previously pressing the "super" keys (e.g. Windows-key) inserted '?'
into the buffer. Now LSUPER/RSUPER/COMPOSE keys will be ignored.